backend/requirements.txt
# Base dependencies
pip>=23.0
setuptools>=65.5.0
wheel>=0.40.0

# Web framework
fastapi>=0.95.2
uvicorn>=0.22.0
sqlalchemy>=2.0.15
pydantic>=1.10.8
python-dotenv>=1.0.0
requests>=2.31.0
email-validator>=2.0.0

# Audio processing
librosa>=0.10.0
pydub>=0.25.1
numpy>=1.23.5
ffmpeg-python>=0.2.0
musicbrainzngs>=0.7.1
pyacoustid>=1.2.0
av>=10.0.0

# Database and caching
psycopg2-binary>=2.9.9
redis>=5.0.1
alembic>=1.13.1

# Authentication and security
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Machine Learning
tensorflow>=2.12.0
pandas>=2.0.0
openpyxl>=3.1.0

# Networking and monitoring
websockets>=12.0
aiohttp>=3.9.1
prometheus-client>=0.19.0
typing-extensions>=4.5.0
psutil>=5.9.0

# Redis
aioredis>=2.0.0
redis>=4.5.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-redis==3.1.2
pytest-benchmark==4.0.0
aioresponses==0.7.4  # For mocking aiohttp responses
